feat: Create `infra/controller.js` and adjust `migrations` and `status` controllers
